# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlabX API",
    "version": "0.1.0",
    "description": "Versioned API contract for the SlabX collectible card marketplace."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/me/addresses/{addressId}": {
      "parameters": [
        {
          "name": "addressId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "patch": {
        "operationId": "updateAddress",
        "summary": "Update an owned address",
        "responses": {
          "200": {
            "description": "Update an owned address"
          }
        }
      },
      "delete": {
        "operationId": "deleteAddress",
        "summary": "Archive an owned address",
        "responses": {
          "204": {
            "description": "Archive an owned address"
          }
        }
      }
    }
  }
}